requirements.txt
python-telegram-bot==21.6
APScheduler==3.6.3
cachetools==5.3.3
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1
certifi==2025.11.12
tornado==6.3.3
six==1.17.0
python-dotenv==1.0.1
